feat: remove dashboard, create home acc usertype
